feat: Implement core avatar upload/retrieval functionality (#15)
- Added MinIO configuration and client setup
- Implemented avatar upload endpoint with size validation and magic number validation
- Added avatar retrieval endpoint with caching headers
- Added error types for object storage operations
- Integrated MinIO service in docker-compose

+++ src/models/defaults.rs
@@ -30,7 +30,7 @@ pub fn default_session_expiration_secs() -> u64 {
     86400
 }
 
-pub fn default_session_cleanup_interval_secs() -> u64 {
+pub fn default_session_cleanup_interval_secs() -> u16 {
     600
 }
